[New] Created test cases for pages-admin-form-builder-action-buttons component from enzyme to rtl
